refactor: restructure Message into message/ subpackage with abstract from_payload_fields
Split the monolithic message.py into message/_message.py (base Message
class with abstract from_payload_fields), message/_metadata.py
(MessageMetadata), and message/__init__.py (public re-exports).

Make from_payload_fields @abstractmethod on Command, Event, and Query
base classes. Update @event_dataclass / @command_dataclass /
@query_dataclass decorators to auto-patch the new method name.
