+++ src/forging_blocks/foundation/messages/message/__init__.py
@@ -1,0 +1,6 @@
+"""Message package — base Message class and MessageMetadata."""
+
+from ._message import Message
+from ._metadata import MessageMetadata
+
+__all__ = ["Message", "MessageMetadata"]
